created routes and sql migration

+++ internal/domain/dto/food_dto.go
@@ -8,8 +8,8 @@ type FoodDetailViewModel struct {
 }
 
 type FoodViewModel struct {
-	ID          int64  `json:"id"`
-	UserID      int64  `json:"user_id"`
+	ID          int    `json:"id"`
+	UserID      int    `json:"user_id"`
 	Title       string `json:"title"`
 	Description string `json:"description"`
 	FoodImage   string `json:"food_image"`
