+++ internal/domain/entity/food.go
@@ -1,8 +1,8 @@
 package entity
 
 type Food struct {
-	ID          int64  `json:"id"`
-	UserID      int64  `json:"user_id"`
+	ID          int    `json:"id"`
+	UserID      int    `json:"user_id"`
 	Title       string `json:"title"`
 	Description string `json:"description"`
 	FoodImage   string `json:"food_image"`
